pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-mcq-generator"
version = "0.1.0"
description = "AI-powered quiz platform with Streamlit"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.108.0",
    "uvicorn>=0.25.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
]